+++ pcs/cli/common/test/test_console_report.py
@@ -3054,3 +3054,18 @@ class SbdWatchdogTestError(NameBuildTest):
             "Unable to initialize test of the watchdog: some reason",
             reports.sbd_watchdog_test_error("some reason"),
         )
+
+
+class ResourceBundleUnsupportedContainerType(NameBuildTest):
+    code = codes.RESOURCE_BUNDLE_UNSUPPORTED_CONTAINER_TYPE
+    def test_success(self):
+        self.assert_message_from_report(
+            (
+                "Bundle 'bundle id' uses unsupported container type, therefore "
+                "it is not possible to set its container options. Supported "
+                "container types are: 'a', 'b', 'c'"
+            ),
+            reports.resource_bundle_unsupported_container_type(
+                "bundle id", ["b", "a", "c"]
+            ),
+        )