+++ pcs/lib/reports.py
@@ -3533,3 +3533,15 @@ def system_will_reset():
     return ReportItem.info(
         report_codes.SYSTEM_WILL_RESET,
     )
+
+
+def resource_bundle_unsupported_container_type(
+    bundle_id, supported_container_types
+):
+    return ReportItem.error(
+        report_codes.RESOURCE_BUNDLE_UNSUPPORTED_CONTAINER_TYPE,
+        info=dict(
+            bundle_id=bundle_id,
+            supported_container_types=sorted(supported_container_types),
+        ),
+    )